tests/wpt/css-tests/css-writing-modes-3_dev/xhtml1/logicalprops-block-size-vlr.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8" />
<title>CSS Logical Properties: {max-,min-}block-size vertical-lr</title>
<link href="mailto:openxu@gmail.com" rel="author" title="Xu Xing" />
<link href="https://drafts.csswg.org/css-logical-props-1/#logical-dimension-properties" rel="help" />
<link href="https://drafts.csswg.org/css-writing-modes-3/#logical-to-physical" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/style-check.js"></script>

<style>
div {
  border: 1px solid #000;
  writing-mode: vertical-lr;
}
#div1 {
  block-size: 40px;
  min-block-size: 50px;
  max-block-size: 100px;
}
#div2 {
  block-size: 100px;
  min-block-size: 50px;
  max-block-size: 100px;
}
#div3 {
  block-size: 120px;
  min-block-size: 50px;
  max-block-size: 100px;
}
#ref_div1 {
  width: 40px;
  min-width: 50px;
  max-width: 100px;
}
#ref_div2 {
  width: 100px;
  min-width: 50px;
  max-width: 100px;
}
#ref_div3 {
  width: 120px;
  min-width: 50px;
  max-width: 100px;
}

p {
  border: 1px solid #000;
  writing-mode: vertical-lr;
}
#p1 {
  block-size: 100px;
  width: 50px;
}
#p2 {
  width: 50px;
  block-size: 100px;
}
#ref_p1 {
  width: 50px;
}
#ref_p2 {
  width: 100px;
}

.table {
  border: 1px solid #000;
  display: table;
  writing-mode: vertical-lr;
}
.tablecell {
  display: table-cell;
  writing-mode: vertical-lr;
}
#table1_cell {
  block-size: 40px;
  min-block-size: 50px;
  max-block-size: 100px;
  inline-size: 100px;
  background-color: red;
}
#table2_cell {
  block-size: 100px;
  min-block-size: 50px;
  max-block-size: 100px;
  inline-size: 100px;
  background-color: blue;
}
#table3_cell {
  block-size: 120px;
  min-block-size: 50px;
  max-block-size: 100px;
  inline-size: 100px;
  background-color: green;
}
#ref_table1_cell {
  width: 40px;
  min-width: 50px;
  max-width: 100px;
  height: 100px;
  background-color: red;
}
#ref_table2_cell {
  width: 100px;
  min-width: 50px;
  max-width: 100px;
  height: 100px;
  background-color: blue;
}
#ref_table3_cell {
  width: 120px;
  min-width: 50px;
  max-width: 100px;
  height: 100px;
  background-color: green;
}
</style>

</head><body><div id="div1"></div>
<div id="div2"></div>
<div id="div3"></div>
<div id="ref_div1"></div>
<div id="ref_div2"></div>
<div id="ref_div3"></div>

<p id="p1">
</p><p id="p2">
</p><p id="ref_p1">
</p><p id="ref_p2">

</p><div class="table">
  <div id="table1_cell" class="tablecell"></div>
</div>
<div class="table">
  <div id="table2_cell" class="tablecell"></div>
</div>
<div class="table">
  <div id="table3_cell" class="tablecell"></div>
</div>
<div class="table">
  <div id="ref_table1_cell" class="tablecell"></div>
</div>
<div class="table">
  <div id="ref_table2_cell" class="tablecell"></div>
</div>
<div class="table">
  <div id="ref_table3_cell" class="tablecell"></div>
</div>

<script>
test(function () {
  assert_true(compareWidthHeight("div1", "ref_div1"));
  assert_true(compareWidthHeight("div2", "ref_div2"));
  assert_true(compareWidthHeight("div3", "ref_div3"));
}, "Check that block-size &lt; min-block-size or min-block-size &lt; block-size &lt;= max-block-size or block-size &gt; max-block-size in vertical-lr");

test(function () {
  assert_true(compareWidthHeight("p1", "ref_p1"));
  assert_true(compareWidthHeight("p2", "ref_p2"));
}, "Check that width override block-size and vice versa in vertical-lr");

test(function () {
  assert_true(compareWidthHeight("table1_cell", "ref_table1_cell"));
  assert_true(compareWidthHeight("table2_cell", "ref_table2_cell"));
  assert_true(compareWidthHeight("table3_cell", "ref_table3_cell"));
}, "Check that block-size &lt; min-block-size or min-block-size &lt; block-size &lt;= max-block-size or block-size &gt; max-block-size in table vertical-lr");
</script>
</body></html>
